+++ vscode/src/completions/inline-completion-item-provider.ts
@@ -45,10 +45,17 @@ export interface CodyCompletionItemProviderConfig {
     featureFlagProvider: FeatureFlagProvider
 }
 
+interface CompletionRequest {
+    document: vscode.TextDocument
+    position: vscode.Position
+    context: vscode.InlineCompletionContext
+}
+
 export class InlineCompletionItemProvider implements vscode.InlineCompletionItemProvider {
     private promptChars: number
     private maxPrefixChars: number
     private maxSuffixChars: number
+    private lastCompletionRequest: CompletionRequest | null = null
     // private reportedErrorMessages: Map<string, number> = new Map()
     private resetRateLimitErrorsAfter: number | null = null
 
@@ -127,6 +134,11 @@ export class InlineCompletionItemProvider implements vscode.InlineCompletionItem
         // Making it optional here to execute multiple suggestion in parallel from the CLI script.
         token?: vscode.CancellationToken
     ): Promise<AutocompleteResult | null> {
+        // Update the last request
+        const lastCompletionRequest = this.lastCompletionRequest
+        const completionRequest: CompletionRequest = { document, position, context }
+        this.lastCompletionRequest = completionRequest
+
         const start = performance.now()
         // We start the request early so that we have a high chance of getting a response before we
         // need it.
@@ -160,6 +172,17 @@ export class InlineCompletionItemProvider implements vscode.InlineCompletionItem
             return null
         }
 
+        let takeSuggestWidgetSelectionIntoAccount = false
+        // Only take the completion widget selection into account if the selection was actively changed
+        // by the user
+        if (
+            this.config.completeSuggestWidgetSelection &&
+            lastCompletionRequest &&
+            onlyCompletionWidgetSelectionChanged(lastCompletionRequest, completionRequest)
+        ) {
+            takeSuggestWidgetSelectionIntoAccount = true
+        }
+
         const docContext = getCurrentDocContext({
             document,
             position,
@@ -167,7 +190,7 @@ export class InlineCompletionItemProvider implements vscode.InlineCompletionItem
             maxSuffixLength: this.maxSuffixChars,
             enableExtendedTriggers: this.config.providerConfig.enableExtendedMultilineTriggers,
             // We ignore the current context selection if completeSuggestWidgetSelection is not enabled
-            context: this.config.completeSuggestWidgetSelection ? context : undefined,
+            context: takeSuggestWidgetSelectionIntoAccount ? context : undefined,
         })
 
         const isIncreasedDebounceTimeEnabled = await this.config.featureFlagProvider.evaluateFeatureFlag(
@@ -267,7 +290,8 @@ export class InlineCompletionItemProvider implements vscode.InlineCompletionItem
                 document,
                 position,
                 result.items,
-                context
+                context,
+                takeSuggestWidgetSelectionIntoAccount
             )
 
             // A completion that won't be visible in VS Code will not be returned and not be logged.
@@ -275,9 +299,10 @@ export class InlineCompletionItemProvider implements vscode.InlineCompletionItem
                 !isCompletionVisible(
                     items,
                     document,
+                    position,
                     docContext,
                     context,
-                    this.config.completeSuggestWidgetSelection,
+                    takeSuggestWidgetSelectionIntoAccount,
                     abortController.signal
                 )
             ) {
@@ -354,7 +379,8 @@ export class InlineCompletionItemProvider implements vscode.InlineCompletionItem
         document: vscode.TextDocument,
         position: vscode.Position,
         items: InlineCompletionItemWithAnalytics[],
-        context: vscode.InlineCompletionContext
+        context: vscode.InlineCompletionContext,
+        takeSuggestWidgetSelectionIntoAccount: boolean = false
     ): vscode.InlineCompletionItem[] {
         return items.map(completion => {
             const currentLine = document.lineAt(position)
@@ -363,7 +389,7 @@ export class InlineCompletionItemProvider implements vscode.InlineCompletionItem
 
             // Append any eventual inline completion context item to the prefix if
             // completeSuggestWidgetSelection is enabled.
-            if (this.config.completeSuggestWidgetSelection && context.selectedCompletionInfo) {
+            if (takeSuggestWidgetSelectionIntoAccount && context.selectedCompletionInfo) {
                 const { range, text } = context.selectedCompletionInfo
                 insertText = text.slice(position.character - range.start.character) + insertText
             }
@@ -456,6 +482,7 @@ function createTracerForInvocation(tracer: ProvideInlineCompletionItemsTracer): 
 function isCompletionVisible(
     completions: vscode.InlineCompletionItem[],
     document: vscode.TextDocument,
+    position: vscode.Position,
     docContext: DocumentContext,
     context: vscode.InlineCompletionContext,
     completeSuggestWidgetSelection: boolean,
@@ -482,7 +509,7 @@ function isCompletionVisible(
     const isAborted = abortSignal ? abortSignal.aborted : false
     const isMatchingPopupItem = completeSuggestWidgetSelection
         ? true
-        : completionMatchesPopupItem(completions, document, context)
+        : completionMatchesPopupItem(completions, position, document, context)
     const isMatchingSuffix = completionMatchesSuffix(completions, docContext)
     const isVisible = !isAborted && isMatchingPopupItem && isMatchingSuffix
 
@@ -522,18 +549,27 @@ function currentEditorContentMatchesPopupItem(
 // VS Code won't show a completion if it won't.
 function completionMatchesPopupItem(
     completions: vscode.InlineCompletionItem[],
+    position: vscode.Position,
     document: vscode.TextDocument,
     context: vscode.InlineCompletionContext
 ): boolean {
     if (context.selectedCompletionInfo) {
         const currentText = document.getText(context.selectedCompletionInfo.range)
         const selectedText = context.selectedCompletionInfo.text
+
         if (completions.length > 0) {
             const visibleCompletion = completions[0]
-            if (
-                typeof visibleCompletion.insertText === 'string' &&
-                !(currentText + visibleCompletion.insertText).startsWith(selectedText)
-            ) {
+            const insertText = visibleCompletion.insertText
+            if (typeof insertText !== 'string') {
+                return true
+            }
+
+            // To ensure a good experience, the VS Code insertion might have the range start at the
+            // beginning of the line. When this happens, the insertText needs to be adjusted to only
+            // contain the insertion after the current position.
+            const offset = position.character - (visibleCompletion.range?.start.character ?? position.character)
+            const correctInsertText = insertText.slice(offset)
+            if (!(currentText + correctInsertText).startsWith(selectedText)) {
                 return false
             }
         }
@@ -562,4 +598,35 @@ function completionMatchesSuffix(completions: vscode.InlineCompletionItem[], doc
     }
 
     return false
+}
+
+/**
+ * Returns true if the only difference between the two requests is the selected completions info
+ * item from the completions widget.
+ */
+function onlyCompletionWidgetSelectionChanged(prev: CompletionRequest, next: CompletionRequest): boolean {
+    if (prev.document.uri.toString() !== next.document.uri.toString()) {
+        return false
+    }
+
+    if (!prev.position.isEqual(next.position)) {
+        return false
+    }
+
+    if (prev.context.triggerKind !== next.context.triggerKind) {
+        return false
+    }
+
+    const prevSelectedCompletionInfo = prev.context.selectedCompletionInfo
+    const nextSelectedCompletionInfo = next.context.selectedCompletionInfo
+
+    if (!prevSelectedCompletionInfo || !nextSelectedCompletionInfo) {
+        return false
+    }
+
+    if (!prevSelectedCompletionInfo.range.isEqual(nextSelectedCompletionInfo.range)) {
+        return false
+    }
+
+    return prevSelectedCompletionInfo.text !== nextSelectedCompletionInfo.text
 }